requirements.txt
# GUI
PyQt5==5.15.10

# OSC Communication (for SuperCollider)
python-osc==1.8.3

# MIDI (for later when we add MIDIMix)
python-rtmidi==1.5.8

# Configuration
pyyaml==6.0.1

# Imaginarium
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
librosa>=0.10.0

# Development
pytest==7.4.3